Conflicts:
      Gemfile
      fake_braintree.gemspec
      lib/fake_braintree.rb
      lib/fake_braintree/credit_card.rb
      lib/fake_braintree/server.rb
      lib/fake_braintree/sinatra_app.rb
      spec/fake_braintree/credit_card_spec.rb

+++ lib/fake_braintree/customer.rb
@@ -1,3 +1,7 @@
+require 'fake_braintree/helpers'
+require 'fake_braintree/credit_card'
+require 'fake_braintree/valid_credit_cards'
+
 module FakeBraintree
   class Customer
     include Helpers
@@ -20,10 +24,15 @@ module FakeBraintree
       if invalid?
         response_for_invalid_card
       else
+        load_payment_method!(
+          @customer_hash.delete('payment_method_nonce'),
+          @customer_hash['credit_card'] ||= {}
+        )
         credit_cards = customer_hash['credit_cards']
         create_customer_with(customer_hash)
         credit_cards.each { |card| add_credit_card_to_registry(card) }
         set_default_credit_card credit_cards.first
+        @customer_hash['created_at'] = Time.now
         response_for_created_customer(customer_hash)
       end
     end
@@ -43,8 +52,12 @@ module FakeBraintree
     end
 
     def delete
-      delete_customer_with_id(customer_id)
-      deletion_response
+      if customer_exists_in_registry?
+        delete_customer_with_id(customer_id)
+        deletion_response
+      else
+        response_for_customer_not_found
+      end
     end
 
     private
@@ -54,7 +67,7 @@ module FakeBraintree
     end
 
     def create_customer_with(hash)
-      FakeBraintree.registry.customers[hash['id']] = hash
+      FakeBraintree.registry.customers[hash['id'].to_s] = hash
     end
 
     def add_credit_card_to_registry(new_credit_card_hash)
@@ -114,6 +127,11 @@ module FakeBraintree
 
     def generate_credit_cards_from(new_credit_card_hash)
       if new_credit_card_hash.present? && new_credit_card_hash.is_a?(Hash)
+        load_payment_method!(
+          new_credit_card_hash.delete('payment_method_nonce'),
+          new_credit_card_hash
+        )
+
         new_credit_card_hash['bin'] = new_credit_card_hash['number'][0..5]
         new_credit_card_hash['last_4'] = new_credit_card_hash['number'][-4..-1]
         new_credit_card_hash['token']  = credit_card_token(new_credit_card_hash)
@@ -149,7 +167,7 @@ module FakeBraintree
     end
 
     def delete_customer_with_id(id)
-      FakeBraintree.registry.customers[id] = nil
+      FakeBraintree.registry.customers.delete(id)
     end
 
     def deletion_response
@@ -194,6 +212,12 @@ module FakeBraintree
 
     def credit_card_token(credit_card_hash_without_token)
       md5("#{credit_card_hash_without_token['number']}#{@customer_hash['merchant_id']}")
+    end
+
+    def load_payment_method!(nonce, credit_card_hash)
+      return unless nonce
+      payment_method_hash = FakeBraintree.registry.payment_methods[nonce]
+      credit_card_hash.merge!(payment_method_hash)
     end
   end
 end
